fix: symlinks and sandboxed builds
Previously, the script file was simply symlinked into the Nix store. However, a
change in sandboxed builds has rendered this approach ineffective. The build
process can no longer resolve the symlink to the home directory, as it is
inaccessible within the sandbox environment. Now we use a proper path, so the
script gets copied into the store.